+++ xbmc/cores/dvdplayer/Edl.cpp
@@ -270,8 +270,8 @@ bool CEdl::ReadEdl(const std::string& strMovie, const float fFramesPerSecond)
       continue;
 
     Cut cut;
-    cut.start = iCutStartEnd[0];
-    cut.end = iCutStartEnd[1];
+    cut.start = static_cast<int>(iCutStartEnd[0]);
+    cut.end = static_cast<int>(iCutStartEnd[1]);
 
     switch (iAction)
     {
@@ -386,8 +386,8 @@ bool CEdl::ReadComskip(const std::string& strMovie, const float fFramesPerSecond
     if (sscanf(szBuffer, "%lf %lf", &dStartFrame, &dEndFrame) == 2)
     {
       Cut cut;
-      cut.start = (int64_t)(dStartFrame / fFrameRate * 1000);
-      cut.end = (int64_t)(dEndFrame / fFrameRate * 1000);
+      cut.start = static_cast<int>(dStartFrame / fFrameRate * 1000);
+      cut.end = static_cast<int>(dEndFrame / fFrameRate * 1000);
       cut.action = COMM_BREAK;
       bValid = AddCut(cut);
     }
@@ -463,8 +463,8 @@ bool CEdl::ReadVideoReDo(const std::string& strMovie)
          *  Times need adjusting by 1/10,000 to get ms.
          */
         Cut cut;
-        cut.start = (int64_t)(dStart / 10000);
-        cut.end = (int64_t)(dEnd / 10000);
+        cut.start = static_cast<int>(dStart / 10000);
+        cut.end = static_cast<int>(dEnd / 10000);
         cut.action = CUT;
         bValid = AddCut(cut);
       }
@@ -476,7 +476,7 @@ bool CEdl::ReadVideoReDo(const std::string& strMovie)
       int iScene;
       double dSceneMarker;
       if (sscanf(szBuffer + strlen(VIDEOREDO_TAG_SCENE), " %i>%lf", &iScene, &dSceneMarker) == 2)
-        bValid = AddSceneMarker((int64_t)(dSceneMarker / 10000)); // Times need adjusting by 1/10,000 to get ms.
+        bValid = AddSceneMarker(static_cast<int>(dSceneMarker / 10000)); // Times need adjusting by 1/10,000 to get ms.
       else
         bValid = false;
     }
@@ -559,8 +559,8 @@ bool CEdl::ReadBeyondTV(const std::string& strMovie)
        * atof() returns 0 if there were any problems and will subsequently be rejected in AddCut().
        */
       Cut cut;
-      cut.start = (int64_t)(atof(pStart->FirstChild()->Value()) / 10000);
-      cut.end = (int64_t)(atof(pEnd->FirstChild()->Value()) / 10000);
+      cut.start = static_cast<int>(atof(pStart->FirstChild()->Value()) / 10000);
+      cut.end = static_cast<int>(atof(pEnd->FirstChild()->Value()) / 10000);
       cut.action = COMM_BREAK;
       bValid = AddCut(cut);
     }
@@ -614,8 +614,8 @@ bool CEdl::ReadPvr(const std::string &strMovie)
   for (it = edl.begin(); it != edl.end(); ++it)
   {
     Cut cut;
-    cut.start = it->start;
-    cut.end = it->end;
+    cut.start = static_cast<int>(it->start);
+    cut.end = static_cast<int>(it->end);
 
     switch (it->type)
     {
